+++ OzzTradeDiary.NET/OzzTradeDiary.WPF/Commands/Exchanges/ExchangeCreateCommand.cs
@@ -19,7 +19,7 @@ namespace TD.WPF.Commands.Exchanges
         protected readonly AbstractDiaryVM _viewModel;
         private readonly IWindowDialogService _windowDialogService;
         private readonly ICurrencyLookupService _currencyLookupService;
-        
+
         public ExchangeCreateCommand(AbstractDiaryVM viewModel, IWindowDialogService windowDialogService,
                                      ICurrencyLookupService currencyLookupService)
         {